Вынести интерфейс ProductService
